Tie label vectors to the lifetime of DigitData
This will be required when we train a kernel perceptron inside
iterDigits and later want to evaluate it on some x vector. Since kernel
perceptrons use phi and labels in evaluation, and the labels are only
valid for the lifetime of the iterDigits lambda, they would be accessing
invalid memory.